+++ src/commands/quickCommand.steps.ts
@@ -5,6 +5,7 @@ import { BranchSorting, configuration, TagSorting } from '../configuration';
 import { Container } from '../container';
 import { GlyphChars, quickPickTitleMaxChars } from '../constants';
 import {
+	BranchSortOptions,
 	GitBranch,
 	GitBranchReference,
 	GitContributor,
@@ -23,6 +24,7 @@ import {
 	RemoteResourceType,
 	Repository,
 	SearchPattern,
+	TagSortOptions,
 } from '../git/git';
 import { GitService } from '../git/gitService';
 import {
@@ -107,16 +109,24 @@ export function appendReposToTitle<
 
 export async function getBranches(
 	repos: Repository | Repository[],
-	options: { filterBranches?: (b: GitBranch) => boolean; picked?: string | string[] } = {},
+	options: { filter?: (b: GitBranch) => boolean; picked?: string | string[]; sort?: BranchSortOptions } = {},
 ): Promise<BranchQuickPickItem[]> {
-	return getBranchesAndOrTags(repos, ['branches'], { sort: true, ...options }) as Promise<BranchQuickPickItem[]>;
+	return getBranchesAndOrTags(repos, ['branches'], {
+		filter: options?.filter != null ? { branches: options.filter } : undefined,
+		picked: options?.picked,
+		sort: options?.sort != null ? { branches: options.sort } : true,
+	}) as Promise<BranchQuickPickItem[]>;
 }
 
 export async function getTags(
 	repos: Repository | Repository[],
-	options: { filterTags?: (t: GitTag) => boolean; picked?: string | string[] } = {},
+	options?: { filter?: (t: GitTag) => boolean; picked?: string | string[]; sort?: TagSortOptions },
 ): Promise<TagQuickPickItem[]> {
-	return getBranchesAndOrTags(repos, ['tags'], { sort: true, ...options }) as Promise<TagQuickPickItem[]>;
+	return getBranchesAndOrTags(repos, ['tags'], {
+		filter: options?.filter != null ? { tags: options.filter } : undefined,
+		picked: options?.picked,
+		sort: options?.sort != null ? { tags: options.sort } : true,
+	}) as Promise<TagQuickPickItem[]>;
 }
 
 export async function getBranchesAndOrTags(
@@ -129,7 +139,7 @@ export async function getBranchesAndOrTags(
 	}: {
 		filter?: { branches?: (b: GitBranch) => boolean; tags?: (t: GitTag) => boolean };
 		picked?: string | string[];
-		sort?: boolean | { branches?: { current?: boolean; orderBy?: BranchSorting }; tags?: { orderBy?: TagSorting } };
+		sort?: boolean | { branches?: BranchSortOptions; tags?: TagSortOptions };
 	} = {},
 ): Promise<(BranchQuickPickItem | TagQuickPickItem)[]> {
 	let branches: GitBranch[] | undefined;
@@ -384,19 +394,19 @@ export async function* pickBranchStep<
 	state: State,
 	context: Context,
 	{
-		filterBranches,
+		filter,
 		picked,
 		placeholder,
 		titleContext,
 	}: {
-		filterBranches?: (b: GitBranch) => boolean;
+		filter?: (b: GitBranch) => boolean;
 		picked?: string | string[];
 		placeholder: string;
 		titleContext?: string;
 	},
 ): AsyncStepResultGenerator<GitBranchReference> {
 	const branches = await getBranches(state.repo, {
-		filterBranches: filterBranches,
+		filter: filter,
 		picked: picked,
 	});
 
@@ -448,20 +458,23 @@ export async function* pickBranchesStep<
 	state: State,
 	context: Context,
 	{
-		filterBranches,
+		filter,
 		picked,
 		placeholder,
+		sort,
 		titleContext,
 	}: {
-		filterBranches?: (b: GitBranch) => boolean;
+		filter?: (b: GitBranch) => boolean;
 		picked?: string | string[];
 		placeholder: string;
+		sort?: BranchSortOptions;
 		titleContext?: string;
 	},
 ): AsyncStepResultGenerator<GitBranchReference[]> {
 	const branches = await getBranches(state.repo, {
-		filterBranches: filterBranches,
+		filter: filter,
 		picked: picked,
+		sort: sort,
 	});
 
 	const step = QuickCommand.createPickStep<BranchQuickPickItem>({
@@ -1254,19 +1267,19 @@ export async function* pickTagsStep<
 	state: State,
 	context: Context,
 	{
-		filterTags,
+		filter,
 		picked,
 		placeholder,
 		titleContext,
 	}: {
-		filterTags?: (b: GitTag) => boolean;
+		filter?: (b: GitTag) => boolean;
 		picked?: string | string[];
 		placeholder: string;
 		titleContext?: string;
 	},
 ): AsyncStepResultGenerator<GitTagReference[]> {
 	const tags = await getTags(state.repo, {
-		filterTags: filterTags,
+		filter: filter,
 		picked: picked,
 	});